rename: Package and file Setting
