+++ src/main/java/com/project/manageus/controller/all/MainController.java
@@ -1,7 +1,6 @@
 package com.project.manageus.controller.all;
 
-import com.project.manageus.bean.UserDTO;
-import com.project.manageus.service.inter.UserService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,11 +8,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class MainController {
-
-
-    @Autowired
-    private UserService userService;
-
 
     @GetMapping("/")
     public String main() {
@@ -36,9 +30,7 @@ public class MainController {
     }
 
     @PostMapping("/register")
-    public String registerPro(UserDTO dto) {
-
-        userService.joinProcess(dto);
+    public String registerPro() {
 
         String url = "redirect:/login";
         return url;
