+++ src/main/java/com/project/manageus/controller/company/UserController.java
@@ -1,4 +1,4 @@
-package com.project.manageus.controller.feature;
+package com.project.manageus.controller.company;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,7 +10,7 @@ public class UserController {
 
     @GetMapping("/")
     public String main() {
-        String url = "/feature/main.html";
+        String url = "/company/main.html";
         return url;
     }
