+++ src/main/java/com/project/manageus/dto/UserDetailDTO.java
@@ -1,0 +1,4 @@
+package com.project.manageus.dto;
+
+public class UserDetailDTO {
+}
