+++ src/main/java/com/project/manageus/entity/UserDetailDTO.java
@@ -1,0 +1,6 @@
+package com.project.manageus.entity;
+
+public class UserDetailDTO {
+
+
+}
