+++ src/main/java/com/project/manageus/repository/UserRepository.java
@@ -4,4 +4,6 @@ import com.project.manageus.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+
+
 }
