+++ src/main/java/com/project/manageus/service/UserService.java
@@ -1,0 +1,10 @@
+package com.project.manageus.service;
+
+
+import com.project.manageus.dto.UserDTO;
+
+public interface UserService {
+
+
+    public void joinProcess(UserDTO dto);
+}
