+++ src/main/java/com/project/manageus/service/UserServiceImpl.java
@@ -1,15 +1,12 @@
-package com.project.manageus.service.impl;
+package com.project.manageus.service;
 
-import com.project.manageus.bean.UserDTO;
-import com.project.manageus.entity.UserEntity;
 import com.project.manageus.repository.UserRepository;
-import com.project.manageus.service.inter.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 @Service
-public class UserServiceImpl implements UserService {
+public class UserServiceImpl {
 
     @Autowired
     private UserRepository userRepository;
@@ -17,8 +14,5 @@ public class UserServiceImpl implements UserService {
     @Autowired
     private BCryptPasswordEncoder bCryptPasswordEncoder;
 
-    @Override
-    public void joinProcess(UserDTO dto) {
 
-    }
 }
