--- src/main/java/com/project/manageus/service/inter/UserService.java
@@ -1,9 +1,0 @@
-package com.project.manageus.service.inter;
-
-import com.project.manageus.bean.UserDTO;
-
-public interface UserService {
-
-
-    public void joinProcess(UserDTO dto);
-}
